Create 'auth' blueprint
See #4
