+++ loginapp/__init__.py
@@ -16,3 +16,7 @@ def create_app(configname):
     db.init_app(app)
 
     # register blueprints
+    from .auth import auth
+    app.register_blueprint(auth)
+
+    return app
