+++ loginapp/auth/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+
+auth = Blueprint("auth", __name__)
+
+from . import views, forms
